// src/player/interfaces/player-batting-season.interface.ts
export interface PlayerBattingSeason {
    year: number;
    teamName: string;
    games: number;
    atBats: number;
    hits: number;
    battingAverage: number;
    RBIs: number;
    baseOnBalls: number;
    hittingStats: string;
    strikeouts: number;
    runs: number;
    stolenBases: number;
    hitByPitch: number;
    sacFly: number;
}