// src/player/interfaces/player-pitching-season.interface.ts
export interface PlayerPitchingSeason {
    year: number;
    teamName: string;
    wins: number;
    losses: number;
    games: string;
    completeGames: number;
    shutouts: number;
    saves: number;
    hitsAllowed: number;
    homerunsAllowed: number;
    earnedRuns: number;
    ERA: number;
    walks: number;
    strikeouts: number;
    opponentAverage: number;
}